Improved documentation. Locked installation to latest software versions at the moment.

+++ install.sh
@@ -2,6 +2,11 @@
 
 INSTALL_DIR=/etc/docker/compose/mlflow
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
 rm -f /etc/systemd/system/mlflow.service
 cp mlflow.service /etc/systemd/system/mlflow.service
 
@@ -13,10 +18,13 @@ cp mlflow.Dockerfile ${INSTALL_DIR}
 cp README.md ${INSTALL_DIR}
 cp requirements.txt ${INSTALL_DIR}
 
-systemctl daemon-reload
-
 echo "MLflow service successfully installed!"
 echo "You can now delete this installation folder including the .env file."
-echo "Use 'systemctl start mlflow' to start MLflow"
-echo "It'll take a while to load the service as the system will pull and build images"
-echo "Use 'systemctl status mlflow' to see its status"
+
+systemctl daemon-reload
+systemctl enable mlflow
+
+echo "I'll start MLflow now."
+systemctl start mlflow
+echo "It'll take a couple of minutes to load the service as the system will pull and build necesary docker images."
+echo "Use 'journalctl -fu mlflow' to see MLflow's installation status."
